Fix sample with indices, ensure return size is equal to input size or indices size
