+++ ding/worker/buffer/utils/__init__.py
@@ -1,0 +1,1 @@
+from .fast_copy import FastCopy, fastcopy
